Create separate Axis component, set up r axis
This commit pulls the Axis class code out of the GameMap class file and
puts it in it's own file.

This commit also sets up and draws the r axis.
